remove all content and only add a basic screen
