--- ExpoImplementationEAS/src/app/index.d.ts
@@ -1,4 +1,0 @@
-declare module '*.png';
-declare module '*.svg';
-declare module '*.jpeg';
-declare module '*.jpg';
