--- ExpoImplementationEAS/src/app/navigation/Screens.ts
@@ -1,5 +1,0 @@
-export enum Screens {
-  HOME = 'Home',
-  SCREEN_VIEWS = 'ScreenViews',
-  PRIVACY = 'Privacy'
-}
