--- ExpoImplementationEAS/src/app/navigation/types.ts
@@ -1,18 +1,0 @@
-/**
- * Learn more about using TypeScript with React Navigation:
- * https://reactnavigation.org/docs/typescript/
- */
-
-import { NativeStackNavigationProp } from '@react-navigation/native-stack';
-import { Screens } from './Screens';
-
-export type RootStackParamList = {
-  [Screens.HOME]: undefined;
-  [Screens.SCREEN_VIEWS]: undefined;
-  [Screens.PRIVACY]: undefined;
-};
-
-export type RootNavigatorNavigationProps =
-  NativeStackNavigationProp<RootStackParamList>;
-
-export type ScreenNamesList = keyof RootStackParamList;
