--- ExpoImplementationEAS/src/app/navigation/useNavigation.ts
@@ -1,5 +1,0 @@
-import { useNavigation as useNavigationBase } from '@react-navigation/native';
-import { RootNavigatorNavigationProps } from './types';
-
-export const useNavigation = (): RootNavigatorNavigationProps =>
-  useNavigationBase<RootNavigatorNavigationProps>();
