--- ExpoImplementationEAS/src/shared/Spacer.tsx
@@ -1,11 +1,0 @@
-import React, { FunctionComponent } from 'react';
-import { View } from 'react-native';
-
-type Props = {
-  height?: number;
-  width?: number;
-};
-
-export const Spacer: FunctionComponent<Props> = ({ height = 0, width = 0 }) => {
-  return <View style={{ height, width }} />;
-};
